plugins/aggregate_raster/1.0/requirements.txt
requests
simplejson
python-dateutil